atomic refactor
